end Avatar, files download style

+++ server/controllers/fileController.js
@@ -159,12 +159,11 @@ class FileController {
     async uploadAvatar(req, res) {
 
         try {
-
             const uid = req.user.id
             const userD = await connection.query(`SELECT * FROM offersworker WHERE id = ${uid}`);
             const user = userD[0][0]
             const avaOld = user.avatar
-            //console.log(avaOld)
+            console.log(avaOld)
             const file = req.files.file
             //console.log(req.files.file)
             // const user = await User.findById(req.user.id)
@@ -173,7 +172,7 @@ class FileController {
 
             user.avatar = avatarName
             await connection.query(`UPDATE offersworker SET avatar = '${avatarName}'   WHERE id = ${uid} `);
-
+            if(avaOld)
             fs.unlinkSync("./files/avatar/" + avaOld)
 
             //await user.save()
@@ -191,18 +190,20 @@ class FileController {
 
     async deleteAvatar(req, res) {
         try {
-
             const uid = req.user.id
             const userD = await connection.query(`SELECT * FROM offersworker WHERE id = ${uid}`);
             const user = userD[0][0]
             const avaOld = user.avatar
 
-            if(avaOld !== ''){
-                fs.unlinkSync("./files/avatar/"+ avaOld)
+            if(avaOld !== '' && avaOld !== null ){
+                if (fs.existsSync("./files/avatar/"+ avaOld)){
+                    fs.unlinkSync("./files/avatar/"+ avaOld)
+                }
+
                 await connection.query(`UPDATE offersworker SET avatar = ''   WHERE id = ${uid} `);
+
             }else{
                 await connection.query(`UPDATE offersworker SET avatar = ''   WHERE id = ${uid} `);
-                console.log('avatar is empty')
             }
 
             user.avatar = null
